feat: add delete article identity validation

+++ controllers/users.js
@@ -1,6 +1,8 @@
 const User = require('../models/user');
 const Article = require('../models/article');
 const NotFoundError = require('../utils/httperrors/notfounderror');
+const AuthorizationError = require('../utils/httperrors/authorizationerror');
+const { UNAUTHORIZED } = require('../utils/httpstatuscodes');
 
 // returns information about the logged-in user (email and name)
 const getCurrentUser = (req, res, next) => {
@@ -16,7 +18,7 @@ const getCurrentUser = (req, res, next) => {
 // returns all articles saved by the user
 const getSavedArticles = (req, res, next) => {
   Article.find({})
-    .orFail(() => new Error('Article list is empty'))
+    .orFail(() => new NotFoundError('Article list is empty'))
     .then((articles) => res.send(articles))
     .catch(next);
 };
@@ -26,16 +28,25 @@ const getSavedArticles = (req, res, next) => {
 const createArticle = (req, res, next) => {
   // const { keyword, title, text, date, source, link, image } = req.body;
   Article.create({ ...req.body, owner: req.user._id })
-    .orFail(next)
     .then((article) => res.send(article))
     .catch(next);
 };
 
 // deletes the stored article by _id
 const deleteArticle = (req, res, next) => {
-  Article.findByIdAndRemove(req.params.articleId)
-    .orFail(() => new NotFoundError('Article not found.'))
-    .then((deletedArticle) => res.send(deletedArticle))
+  Article.findById(req.params.articleId)
+    .select('owner')
+    .orFail(() => next(new NotFoundError('Article not found')))
+    .then((foundArticle) => {
+      // Validates that current user is the owner of the requested article to be deleted.
+      if (req.user._id !== String(foundArticle.owner)) {
+        throw new AuthorizationError("Cannot delete other users' articles.", UNAUTHORIZED);
+      }
+      Article.findByIdAndRemove(req.params.articleId)
+        .orFail(() => new NotFoundError('Article not found.'))
+        .then((deletedArticle) => res.send(deletedArticle))
+        .catch(next);
+    })
     .catch(next);
 };
